pyproject.toml
[project]
name = "productdna-assessment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.3",
    "djangorestframework>=3.16.0",
    "ruff>=0.11.13",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]
